+++ NEWS.md
@@ -14,6 +14,8 @@
   contains a video track". If enabled, only source files containing video
   tracks will be used for setting the destination file name. Other files that
   are added are ignore. Implements the rest of #2058.
+* MKVToolNix GUI: job queue: selected jobs can now be move up and down by
+  pressing the `Ctrl+Up` and `Ctrl+Down` keys. Implements part of #2060.
 
 ## Bug fixes